requirements.txt
requests>=2.28.0
aiofiles>=23.1.0
aiohttp>=3.8.5
aiometer>=0.4.0
gitpython>=3.1.31
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
asyncio>=3.4.3
aiodns>=3.0.0
colorama>=0.4.6